added blueprint for dashboard
