+++ src/__init__.py
@@ -24,6 +24,7 @@ def create_app(config_class=Config):
     from src.models import Role, User, user_role  # noqa
     from src.project.routes import project
     from src.service.routes import service
+    from src.userDashboard.routes import dashboard
     from src.users.routes import users
 
     app.register_blueprint(admin)
@@ -32,6 +33,7 @@ def create_app(config_class=Config):
     app.register_blueprint(project)
     app.register_blueprint(service)
     app.register_blueprint(users)
+    app.register_blueprint(dashboard)
 
     _db.init_app(app)
     _bcrypt.init_app(app)
